update iresearch (#14901)
* update iresearch

* update iresearch

* wip

* turn off stemming for Chinese

* fix tests

+++ 3rdParty/iresearch/core/analysis/text_token_normalizing_stream.cpp
@@ -380,7 +380,7 @@ bool normalizing_token_stream::reset(const string_ref& data) {
   }
 
   // convert input string for use with ICU
-  if (data.size() > std::numeric_limits<int32_t>::max()) {
+  if (data.size() > static_cast<uint32_t>(std::numeric_limits<int32_t>::max())) {
     return false;
   }
